chore: Make TopAndSidebar a Wrapper component
You can now pass children into TopAndSidebar.tsx for the different Home Page
screens
